Merged PR 6356: Added Doc Type Confidence Score to Model and UI
[APH-159](https://mmitnetwork.atlassian.net/browse/APH-159)

## Backend
Added `doc_type_confidence` score that is generated by scrape worker to the document model. Scrape worker now stores this on the document when it is scraped.

## Frontend
Displays `doc_type_confidence` in the UI in the documents table and the document form.

In document form, when a user selects another document type the document confidence is hidden since the confidence percentage no longer applies to the updated doc type.