Merged PR 6484: APH-180: Add permissions to Update Service from SourceHub App
This PR will add permissions to the IAM role for SourceHub App to allow it to UpdateService on other Services in the cluster to programatically modify the `desiredCount` of the service as a manual way of scaling out.

Additionally, in Terraform, declare the `desiredCount` property be ignored when applying changes and also added some missing tags.