Merged PR 6427: APH-165 Added URL Uniqueness Validation
[APH-165](https://mmitnetwork.atlassian.net/browse/APH-165)

URLs should be unique across all Sites

## Backend
Added GET /sites/active-url endpoint that checks if provided URL is already in use on another site (excludes current site).

## Frontend
Added custom validator to URL fields to call the sites/active-url endpoint and reject the input if the url is already in use.