Merged PR 6506: Prevent Site Deletion if Collection Records Are Present
[APH-223](https://mmitnetwork.atlassian.net/browse/APH-223)

##Backend
`DELETE /sites/{id}` endpoint now checks site to be deleted for associated collection records. If collections are present, API returns a 405 response with an error message detailing how many collections were found.

##Frontend
Sites home page now handles the response when deleting a site. On error, an error notification is displayed, and on success, a success notification is displayed