Merged PR 6816: APH-278: Configure Event Bridge
This PR will create an Event Bus in EventBridge, create an IAM policy for putting events on the Event Bus, and assigning that policy to the SourceHub App role. Additionally, EVENT_BUS_ARN and EVENT_BUS_SOURCE are injected into the SourceHub App service as environment vars.

For now, assuming source is `com.mmitnetwork.sourcehub`. This is an AWS recommendation for setting source names.