Merged PR 7202: APH-364: Disable Terraform locks in validation stage
Disable Terraform locks in validation stage so that locks aren't orphaned when the pipeline is cancelled due to new commits being pushed to the PR.

Disabling locks can be dangerous, as there could be another 'apply' operation running at the same time. But I believe this is mitigated because in this step we are only doing plans to validate the Terraform deployment and we aren't retaining the saved plan to be used in a later 'apply' step. Those 'apply' steps will put a lock on the state.