Merged PR 7271: Handle no text_checksum on existing document
[APH-273](https://mmitnetwork.atlassian.net/browse/APH-273)

Many existing documents were created before we saved text and text_checksums. As result, they perpetually have a null text_checksum.

Added check for text_checksum on existing document in scrapeworker. If text_checksum is null, checksum is created and added to document. We may be able to remove this down the line once things are caught up

Added error handling to the diff route for docs with no text_checksum