Merged PR 8918: APH-773-traceback issues to resolve
https://mmitnetwork.atlassian.net/browse/APH-773

for the particular site in the ticket, filename in the downloaded .xlsx response header didn't have a semicolon as a last letter. our regex for content_disposition_filename required semicolon in the end, made it optional. that way the files won't have .none extension which was causing a bug.

for #2, get_extension_from_mimetype() is overwriting None value.