Merged PR 8975: Tag Compare Sections by Text
[APH-722](https://mmitnetwork.atlassian.net/browse/APH-722)

Adjusted document "sectioning" that is done to pair tags. If a key section is provided, we use the text found inside of the key area as the identifier of that section, instead of the tag(s). This allows users to designate sections that are defined by text other than tags.

**Tag Compare**
- Updated tag compare to use key text for section ID if key sections are available.
- Added check to mark added / removed focus tags within a section
- Added corresponding tests