Merged PR 9603: APH-970 Fix Dates not Reseting in Queue
APH-970 Fix Dates not Reseting in Queue

ListDatePicker was maintaining its own internal state, which was not getting reset when user 'Take Next' s in queue because all components remain alive, only new data is swapped in.

Solution is to remove internal state of ListDatePicker and use antd form state management. aka just have ListDatePicker accept a 'value' and 'onChange' prop, then use ListDatePicker in a Form.Item with the 'name' prop to bind it to the correct value.