Merged PR 9913: APH-1044 Only poll for table updates when tab is active
Only poll for table updates when tab is active.

Users can end up with large numbers of SourceHub tabs, each polling pointlessly, refreshing it's site, document, or scrape data.

This modified useInterval to not update watermark when tab is not active. Additionally, when the tab regains focus, the watermark is instantly updated.

Also refactors several uses of 'pollingInterval' to use useInterval to take advantage of this.