Merged PR 9928: APH-1050 Fix partial updates clear date values
APH-1050 Fix partial updates clear date values

When we apply an update, we uses Pydantic BaseModel's ability to differentiate between unset and None to determine whether to clear a field (e.g. update value is None) or leave a field as is (e.g. update value is unset). For dates however, the cast_dates method always sets the value to something, either the cleaned up value or None. This turns any update that does not mention a date into a update that clears the date.

Fix is to only cast the date if it exists.