Merged PR 10166: APH-877 Lineage ML Model
Use an ML Model to construct lineages between documents.

The model provides, for any 2 documents, a confidence score on those documents being in the same lineage.

When a site is collected, all new documents are scores against each other, and any documents with a score > 0.5 are grouped together and sorted by final effective date to form a lineage.

These new lineages are then 'attached' to existing lineages for the site by calculating a score for the last doc of all existing lineages against the first doc of all new lineages.

The model is trained based on a relatively small number of sites, whose lineage could be manually corrected. Because of the low number of sites, there are many patterns the model has not 'seen' and so will not perform well against. In the near future, a list of correctly lineages sites should be recollected and used to improve the model. Additional feature engineering will be required once more data is available.

While the model is untested/poorly trained, all documents require a confidence of 1 (impossible) to skip the classification queue.