Merged PR 10593: APH-1213 enable bcbf florida scraper
we need to click on both dropdowns everytime after we click on new links. this updates the DOM and makes new iframe visible.

after sending a lot of test requests to the page, I started getting constant timeouts, switching the IP helped. updated the code to wait 10 seconds after timeout and 0.25 after each click on a link. this helped downloading all the pdf links.