Merged PR 10701: APH-1250: Model Build Task and Scheduler
This PR defines an ECS Task to build the SourceHub models and push to S3. The task will be started via an EventBridge scheduler using cron scheduling.

NOTE: I needed to bump the AWS Terraform provider version to 4.51.0 to use the EventBridge Scheduler resources. This will cause messages under the section `Objects have changed outside of Terraform` in the Terraform Plan